requirements.txt
aiogram~=3.0.0b7
python-dotenv~=1.0.0
Pillow~=9.5.0
torch~=2.0.0
torchvision~=0.15.1